+++ document/src/main/kotlin/com/ritense/document/exporter/JsonSchemaDocumentDefinitionExporter.kt
@@ -14,37 +14,36 @@
  * limitations under the License.
  */
 
-package com.ritense.document.service
+package com.ritense.document.exporter
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.ritense.document.domain.search.SearchConfigurationDto
+import com.ritense.document.domain.impl.JsonSchemaDocumentDefinitionId
+import com.ritense.document.service.impl.JsonSchemaDocumentDefinitionService
 import com.ritense.exporter.ExportFile
 import com.ritense.exporter.ExportResult
 import com.ritense.exporter.Exporter
 import com.ritense.exporter.request.DocumentDefinitionExportRequest
-import java.io.ByteArrayOutputStream
 import org.springframework.transaction.annotation.Transactional
+import java.io.ByteArrayOutputStream
 
 @Transactional(readOnly = true)
-class SearchFieldExporter(
+class JsonSchemaDocumentDefinitionExporter(
     private val objectMapper: ObjectMapper,
-    private val searchFieldService: SearchFieldService,
+    private val documentDefinitionService: JsonSchemaDocumentDefinitionService
 ) : Exporter<DocumentDefinitionExportRequest> {
 
-    override fun supports() = DocumentDefinitionExportRequest::class.java
+    override fun supports(): Class<DocumentDefinitionExportRequest> =
+        DocumentDefinitionExportRequest::class.java
 
     override fun export(request: DocumentDefinitionExportRequest): ExportResult {
-        val searchFields = searchFieldService.getSearchFields(request.name)
-
-        if (searchFields.isEmpty()) {
-            return ExportResult()
-        }
+        val documentDefinitionId = JsonSchemaDocumentDefinitionId.existingId(request.name, request.version)
+        val documentDefinition = documentDefinitionService.findBy(documentDefinitionId).orElseThrow()
 
         val exportFile = ByteArrayOutputStream().use {
-            objectMapper.writerWithDefaultPrettyPrinter().writeValue(it, SearchConfigurationDto(searchFields))
+            objectMapper.writerWithDefaultPrettyPrinter().writeValue(it, documentDefinition.schema.asJson())
 
             ExportFile(
-                PATH.format(request.name),
+                PATH.format(documentDefinition.id.name()),
                 it.toByteArray()
             )
         }
@@ -53,6 +52,6 @@ class SearchFieldExporter(
     }
 
     companion object {
-        private const val PATH = "config/search/%s.json"
+        private const val PATH = "config/document/definition/%s.schema.json"
     }
 }
