+++ document/src/test/kotlin/com/ritense/document/service/JsonSchemaDocumentDefinitionImporterTest.kt
@@ -16,6 +16,7 @@
 
 package com.ritense.document.service
 
+import com.ritense.document.importer.JsonSchemaDocumentDefinitionImporter
 import com.ritense.document.service.impl.JsonSchemaDocumentDefinitionService
 import com.ritense.importer.ImportRequest
 import org.assertj.core.api.Assertions.assertThat
