+++ document/src/test/kotlin/com/ritense/document/service/SearchFieldExporterIntTest.kt
@@ -18,6 +18,7 @@ package com.ritense.document.service
 
 import com.ritense.authorization.AuthorizationContext.Companion.runWithoutAuthorization
 import com.ritense.document.BaseIntegrationTest
+import com.ritense.document.exporter.SearchFieldExporter
 import com.ritense.exporter.request.DocumentDefinitionExportRequest
 import org.junit.jupiter.api.Test
 import org.skyscreamer.jsonassert.JSONAssert
